Movidos dialogs de tirada generica D6 y D10 al sidebar. Falta arreglar como se visualizan
