+++ templates/actors/jugador.html
@@ -6,13 +6,6 @@
         <img class="profile-img" src="{{img}}" data-edit="img" title="{{name}}"/>
         <h1 class="charname"><input name="name" type="text" value="{{name}}" placeholder="Nombre"/></h1>
       </div>
-      <div>
-        <h3>Tirada de dados</h3>
-        <div class="flexrow">
-          <div class="D10Roll"><i class="fas fa-dice"><span class="tooltiptext"></span></i><label">D10</label></div>
-          <div class="D6Roll"><i class="fas fa-dice"></i><label>D6</label></div>
-        </div>
-      </div>
       <div class="atributos">
         <h3>{{localize 'attributes'}}</h3>
         <div class="flexrow">
